changed from vs project to general mingw

+++ include/RenderWindow.hpp
@@ -10,7 +10,7 @@ public:
 
 	void cleanUp();
 	void clear();
-	void render(Entity& p_entity, int p_x, int p_y, const float& p_cellSize);
+	void render(Entity& p_entity, int p_x, int p_y, const int& p_cellSize);
 	void display();
 
 	const int getWidth() const;
