+++ src/RenderWindow.cpp
@@ -26,7 +26,7 @@ void RenderWindow::clear() {
 	SDL_RenderClear(renderer);
 }
 
-void RenderWindow::render(Entity& p_entity, const int p_x, const int p_y, const float& p_cellSize) {
+void RenderWindow::render(Entity& p_entity, const int p_x, const int p_y, const int& p_cellSize) {
 
 	SDL_Rect rect = { p_x, p_y, p_cellSize, p_cellSize};
